pyproject.toml
[project]
name = "voynich"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "scipy>=1.17.1",
    "python-Levenshtein>=0.26.1",
]
